feat: Added endpoint for roles and permission
