+++ alembic/versions/085de908c797_create_squeeze_table.py
